Remove package name stutter from types defined in runner/execer/os and make processWatcher methods public
