+++ runner/execer/os/easy_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/twitter/scoot/common/log/hooks"
 	"github.com/twitter/scoot/common/log/tags"
 	"github.com/twitter/scoot/common/stats"
-	"github.com/twitter/scoot/runner/execer"
+	scootexecer "github.com/twitter/scoot/runner/execer"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -23,7 +23,7 @@ func TestAll(t *testing.T) {
 	e := NewBoundedExecer(0, nil)
 
 	// TODO(dbentley): factor out an assertRun method
-	cmd := execer.Command{
+	cmd := scootexecer.Command{
 		Argv: []string{"true"},
 		LogTags: tags.LogTags{
 			Tag:    "tag",
@@ -36,17 +36,17 @@ func TestAll(t *testing.T) {
 		t.Fatalf("Couldn't run true %v", err)
 	}
 	status := p.Wait()
-	if status.State != execer.COMPLETE || status.ExitCode != 0 {
+	if status.State != scootexecer.COMPLETE || status.ExitCode != 0 {
 		t.Fatalf("Got unexpected status running true %v", status)
 	}
 
-	cmd = execer.Command{Argv: []string{"false"}}
+	cmd = scootexecer.Command{Argv: []string{"false"}}
 	p, err = e.Exec(cmd)
 	if err != nil {
 		t.Fatalf("Couldn't run false %v", err)
 	}
 	status = p.Wait()
-	if status.State != execer.COMPLETE || status.ExitCode != 1 {
+	if status.State != scootexecer.COMPLETE || status.ExitCode != 1 {
 		t.Fatalf("Got unexpected status running false %v", status)
 	}
 
@@ -59,7 +59,7 @@ func TestOutput(t *testing.T) {
 
 	stdoutExpected := "hello world\n"
 	// TODO(dbentley): factor out an assertRun method
-	cmd := execer.Command{
+	cmd := scootexecer.Command{
 		Argv:   []string{"echo", "-n", stdoutExpected},
 		Stdout: &stdout,
 		Stderr: &stderr,
@@ -74,7 +74,7 @@ func TestOutput(t *testing.T) {
 		t.Fatalf("Couldn't run true %v", err)
 	}
 	status := p.Wait()
-	if status.State != execer.COMPLETE || status.ExitCode != 0 {
+	if status.State != scootexecer.COMPLETE || status.ExitCode != 0 {
 		t.Fatalf("Got unexpected status running true %v", status)
 	}
 	stdoutText, stderrText := stdout.String(), stderr.String()
@@ -88,7 +88,7 @@ func TestMemUsage(t *testing.T) {
 	// Command to increase memory by 1MB every .1s until we hit 50MB after 5s.
 	// Creates a bash process and under that a python process. They should both contribute to MemUsage.
 	str := `import time; exec("x=[]\nfor i in range(50):\n x.append(' ' * 1024*1024)\n time.sleep(.1)")`
-	cmd := execer.Command{
+	cmd := scootexecer.Command{
 		Argv: []string{"python", "-c", str},
 		LogTags: tags.LogTags{
 			Tag:    "tag",
@@ -97,22 +97,22 @@ func TestMemUsage(t *testing.T) {
 		},
 	}
 	e := NewBoundedExecer(0, nil)
-	process, err := e.Exec(cmd)
+	p, err := e.Exec(cmd)
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	process.(*osProcess).ats = 1
-	defer process.Abort()
+	p.(*process).ats = 1
+	defer p.Abort()
 	// Check for growing memory usage at [1.5, 3]s. Then check that the usage is between a reasonable min/max.
 	prevUsage := 0
 	sleepDuration := 500 * time.Millisecond
 	for i := 0; i < 2; i++ {
 		time.Sleep(sleepDuration)
-		err := e.pw.getAndSetProcs()
+		err := e.pw.GetAndSetProcs()
 		if err != nil {
 			t.Fatal(err)
 		}
-		if newUsage, err := e.pw.memUsage(process.(*osProcess).cmd.Process.Pid); err != nil {
+		if newUsage, err := e.pw.MemUsage(p.(*process).cmd.Process.Pid); err != nil {
 			t.Fatalf(err.Error())
 		} else if int(newUsage) <= prevUsage {
 			t.Fatalf("Expected growing memory, got: %d -> %d @%dms", prevUsage, newUsage, (i+1)*int(sleepDuration/time.Millisecond))
@@ -132,8 +132,8 @@ func TestMemCap(t *testing.T) {
 	// Command to increase memory by 1MB every .1s up to 5s.
 	// Creates a bash process and under that a python process. They should both contribute to MemUsage.
 	str := `import time; exec("x=[]\nfor i in range(50):\n x.append(' ' * 1024*1024)\n time.sleep(.1)")`
-	memCh := make(chan execer.ProcessStatus)
-	cmd := execer.Command{
+	memCh := make(chan scootexecer.ProcessStatus)
+	cmd := scootexecer.Command{
 		Argv: []string{"python", "-c", str},
 		LogTags: tags.LogTags{
 			Tag:    "tag",
@@ -143,26 +143,26 @@ func TestMemCap(t *testing.T) {
 		MemCh: memCh,
 	}
 	// Terminate nearly immediately, after memory grows to 1MB.
-	e := NewBoundedExecer(execer.Memory(1024*1024), stats.NilStatsReceiver())
-	process, err := e.Exec(cmd)
+	e := NewBoundedExecer(scootexecer.Memory(1024*1024), stats.NilStatsReceiver())
+	p, err := e.Exec(cmd)
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	process.(*osProcess).ats = 1
-	defer process.Abort()
-	pid := process.(*osProcess).cmd.Process.Pid
-	var usage execer.Memory
+	p.(*process).ats = 1
+	defer p.Abort()
+	pid := p.(*process).cmd.Process.Pid
+	var usage scootexecer.Memory
 	var timeoutCh <-chan time.Time
 	timeout := time.NewTimer(time.Second * 2)
 	timeoutCh = timeout.C
 	defer timeout.Stop()
 	select {
 	case <-memCh:
-		err := e.pw.getAndSetProcs()
+		err := e.pw.GetAndSetProcs()
 		if err != nil {
 			t.Fatal(err)
 		}
-		if usage, err = e.pw.memUsage(pid); err != nil {
+		if usage, err = e.pw.MemUsage(pid); err != nil {
 			// We don't return this error because it just means the process was already killed
 			log.Errorf("Error finding memUsage: %s", err)
 		}
